+++ tests/integration/cli/instance_segmentation/test_instance_segmentation.py
@@ -32,7 +32,15 @@ args = {
     "--val-data-roots": "tests/assets/car_tree_bug",
     "--test-data-roots": "tests/assets/car_tree_bug",
     "--input": "tests/assets/car_tree_bug/images/train",
-    "train_params": ["params", "--learning_parameters.num_iters", "1", "--learning_parameters.batch_size", "2"],
+    "train_params": [
+        "params",
+        "--learning_parameters.num_iters",
+        "1",
+        "--learning_parameters.batch_size",
+        "2",
+        "--postprocessing.max_num_detections",
+        "200",
+    ],
 }
 
 # Training params for resume, num_iters*2